ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
boto3==1.35.36
pydantic==2.9.2
pydantic-settings==2.5.2
